pyproject.toml
[tool.poetry]
name = "autoriaproject"
version = "0.1.0"
description = ""
authors = ["SaLuTeK1 <romanbabiichuk2811@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mysqlclient = "^2.2.4"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django = "^5.0.6"
pillow = "^10.3.0"
django-filter = "^24.2"
drf-yasg = "^1.21.7"
celery = "^5.4.0"
redis = "^5.0.7"
djangochannelsrestframework = "^1.2.0"
daphne = "^4.1.2"
channels-redis = "^4.2.0"
requests = "^2.32.3"
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"

[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
